@@ -15,32 +15,33 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.3",
-    "@angular/animations": "^5.2.11",
-    "@angular/cdk": "^5.2.5",
-    "@angular/common": "^5.2.11",
-    "@angular/compiler": "^5.2.11",
-    "@angular/core": "^5.2.11",
-    "@angular/forms": "^5.2.11",
-    "@angular/http": "^5.2.11",
-    "@angular/material": "^5.2.5",
-    "@angular/platform-browser": "^5.2.11",
-    "@angular/platform-browser-dynamic": "^5.2.11",
-    "@angular/platform-server": "^5.2.11",
-    "@angular/router": "^5.2.11",
+    "@angular-devkit/build-angular": "^0.6.8",
+    "@angular/animations": "^6.0.7",
+    "@angular/cdk": "^6.3.2",
+    "@angular/common": "^6.0.7",
+    "@angular/compiler": "^6.0.7",
+    "@angular/core": "^6.0.7",
+    "@angular/forms": "^6.0.7",
+    "@angular/http": "^6.0.7",
+    "@angular/material": "^6.3.2",
+    "@angular/platform-browser": "^6.0.7",
+    "@angular/platform-browser-dynamic": "^6.0.7",
+    "@angular/platform-server": "^6.0.7",
+    "@angular/router": "^6.0.7",
     "@asymmetrik/ngx-leaflet": "^3.1.0",
-    "agm-direction": "^0.3.5",
+    "agm-direction": "^0.5.9",
     "core-js": "^2.5.7",
     "font-awesome": "^4.7.0",
     "leaflet": "^1.3.1",
     "leaflet-overpass-layer": "^2.8.3",
     "primeng": "^5.2.7",
-    "rxjs": "^5.5.11",
+    "rxjs": "^6.2.1",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.11",
-    "@angular/language-service": "^5.2.11",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "^6.0.7",
+    "@angular/language-service": "^6.0.7",
     "@types/geojson": "^7946.0.4",
     "@types/googlemaps": "^3.30.11",
     "@types/jasmine": "^2.8.8",
@@ -59,6 +60,6 @@
     "protractor": "^5.3.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "^2.7.2"
   }
 }
