@@ -2,7 +2,7 @@ import { HydrantFeature } from './../hydrant-feature';
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { LatLngBounds } from 'leaflet';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { HttpResponse } from 'selenium-webdriver/http';
 
 @Injectable()
